fix: move task 6 functionality to products service stack
