+++ lib/products-service-stack.ts
@@ -1,7 +1,11 @@
 import * as cdk from 'aws-cdk-lib';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
+import * as lambdaEventSource from 'aws-cdk-lib/aws-lambda-event-sources';
 import * as apigw from 'aws-cdk-lib/aws-apigatewayv2';
 import * as iam from 'aws-cdk-lib/aws-iam';
+import * as sqs from 'aws-cdk-lib/aws-sqs';
+import * as sns from 'aws-cdk-lib/aws-sns';
+import * as snsSubscriptions from 'aws-cdk-lib/aws-sns-subscriptions';
 import { HttpLambdaIntegration } from 'aws-cdk-lib/aws-apigatewayv2-integrations';
 import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
 import { Construct } from 'constructs';
@@ -10,6 +14,8 @@ import { ALLOWED_HEADERS, LAYERS_PATH, ORIGINS, SERVICES_PATH, API_PATHS } from 
 import path = require('node:path');
 
 export class ProductsServiceStack extends cdk.Stack {
+  public readonly CatalogItemsQueue: sqs.Queue;
+
   constructor(scope: Construct, id: string, props?: cdk.StackProps) {
     super(scope, id, props);
 
@@ -117,5 +123,57 @@ export class ProductsServiceStack extends cdk.Stack {
       methods: [apigw.HttpMethod.GET],
       integration: getProductByIdIntegration,
     });
+
+    // Create SQS Import Products Queue
+    const catalogItemsQueue = new sqs.Queue(this, 'ImportProductsQueue', {
+      queueName: 'import-products-queue-carp',
+    });
+    this.CatalogItemsQueue = catalogItemsQueue;
+
+    // Create SQS event source for catalog batch process lambda function
+    const catalogBatchProcessEventSource = new lambdaEventSource.SqsEventSource(catalogItemsQueue, {
+      batchSize: 5,
+    });
+
+    // Create SNS import service topic
+    const createProductTopic = new sns.Topic(this, 'ImportTopic');
+
+    // Add SNS import service topic subscriptions
+    createProductTopic.addSubscription(new snsSubscriptions.EmailSubscription(process.env.SNS_SUB_SUCCESS_EMAIL!, {
+      filterPolicy: {
+        status: sns.SubscriptionFilter.stringFilter({
+          allowlist: ['success'],
+        }),
+      }
+    }));
+
+    createProductTopic.addSubscription(new snsSubscriptions.EmailSubscription(process.env.SNS_SUB_FAIL_EMAIL!, {
+      filterPolicy: {
+        status: sns.SubscriptionFilter.stringFilter({
+          allowlist: ['error'],
+        }),
+      },
+    }));
+
+    // Create catalog batch process lambda function
+    const catalogBatchProcess = new NodejsFunction(this, 'CatalogBatchProcessHandler', {
+      runtime: lambda.Runtime.NODEJS_20_X,
+      entry: path.join(SERVICES_PATH.Products, 'catalogBatchProcess.ts'),
+      layers: [utilsLayer, uuidLayer, zodLayer],
+      initialPolicy: [new iam.PolicyStatement({
+        effect: iam.Effect.ALLOW,
+        actions: ['dynamodb:PartiQLInsert'],
+        resources: [DYNAMO_ARNS.products, DYNAMO_ARNS.stocks],
+      })],
+      environment: {
+        SNS_ARN: createProductTopic.topicArn,
+        PRODUCTS_TABLE: TableNames.Products,
+        STOCKS_TABLE: TableNames.Stocks,
+      },
+      events: [catalogBatchProcessEventSource],
+    });
+
+    // Allow catalog batch process lambda function to send messages to the topic
+    createProductTopic.grantPublish(catalogBatchProcess);
   }
 };
